+++ src/main/java/login/tikichat/domain/top_ranked_chatroom/batch/StatsToTopRankedRoomsProcessor.java
@@ -1,5 +1,6 @@
-package login.tikichat.domain.chatroom.scheduler;
+package login.tikichat.domain.top_ranked_chatroom.batch;
 
+import login.tikichat.domain.top_ranked_chatroom.dto.ChatRoomStatsDto;
 import login.tikichat.domain.top_ranked_chatroom.model.TopRankedChatRoom;
 import org.springframework.batch.core.configuration.annotation.StepScope;
 import org.springframework.batch.item.ItemProcessor;
